move seeding files to their own folder
