+++ database/model.js
@@ -1,6 +1,5 @@
 require('dotenv').config();
 const { Description } = require('./database');
-// const { Amenity } = require('./database');
 
 const addDescription = (listing) => {
   new Description({
@@ -20,17 +19,5 @@ const addDescription = (listing) => {
       throw err;
     });
 };
-
-// const addAmenity = (listing) => {
-//   new Amenity({
-//     id: listing.id,
-//     basics: listing.basics,
-//     dining: listing.dining,
-//     guestAccess: listing.guestAccess,
-//     bedAndBath: listing.bedAndBath,
-//     safetyFeatures: listing.safetyFeatures,
-//     notIncluded: listing.notIncluded,
-//   }).save();
-// };
 
 module.exports = { addDescription };
