+++ seedDatabase/amenitiesSeedData.js
